requirements.txt
albumentations==2.0.5
einops==0.8.1
hydra-core==1.3.2
numpy==2.0.2
omegaconf==2.3.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
pytorch_lightning==2.5.0.post0
torchmetrics==1.6.2
safetensors==0.5.3
scikit_learn==1.6.1
scipy==1.15.2
segmentation_models_pytorch==0.4.0
scikit-image==0.25.2
timm==1.0.15
wandb==0.19.10
omegaconf==2.3.0
matplotlib==3.10.1
cellpose==3.1.1.2
seaborn==0.13.2
xgboost==2.1.4
